Fixes for celery remote plugin
We return the task id from the celery task to be used downstream. This
must then be passed on. The transformation status api is extended to
take this id as an argument. This requires a change of the data model
and api and calls. The transformation status will now give the status of
the celery remote job