Implement CI/CD for dependabot
The workflows for dependabot are there to auto-merge the dependabot PRs
that will arise into a separate branch `ci/dependabot-updates`.
Also for creating a single PR to update dependencies and pre-commit
versions once a week - one can consider moving this to a less frequent
timing.
A workflow also updates the separate branch `ci/dependabot-updates` with
the changes from `master` on every push to `master`.