Add test of plugins loader
Note: The current installation includes old names for the parse
strategies for JPEG and JSON. This would cause this strategy to fail,
but the test includes a *temporary* workaround that avoids the issue
and raises warnings rather than exceptions. This workaround should be
removed when the installation has adopted the new names for these
two strategies.