Fix deleting entries in AttrDict
When deleting entries in AttrDict they should be completely removed from
the `__dict__` in case the entry is not a defined pydantic field,
otherwise it should be reset to the default value and removed from the
set of set fields.

Fix `update()` method to check for a pydantic model (after checking for
a `dict` or `Mapping`) and then convert the model to a `dict`.