Update pytest jobs in CI
It seems the "extra libs" part were not actually necessary any longer as
the extra libs (ase and numpy) are already part of the installed
packages.
So that part has been removed, the coverage flags have also been
removed, and the new "testing" extra is used when installing the
package, instead of "dev".