Update release and tests CI/CD workflows
CD - Release workflow:
- Ensure release workflow only runs from 'release/pydantic-v1-support'
branch when doing a v0.5.x release.
- Change release_branch name to 'release/pydantic-v1-support' in order
to update this branch with the commit that the new versioned tag is
pointing to. This will never be part of the 'master' branch.
- Add new input to ensure the 'latest' "version" is not updated in the
documentation. Note, this new input is not yet available, but will be
available in SINTEF/ci-cd > v2.5.1.

CI - Tests:
- Ensure the tests are run upon pushes to 'release/pydantic-v1-support'
to mimic the behaviour of 'master' as best as possible.