Apply suggestions from code review
- Revert removal of urllib3 dependency as this is out-of-scope.
- Add back pydantic configuration doc strings.
- Remove superfluous method doc strings in AttrDict.
- Add DeprecationWarning for AttrDict.__delitem__ to notify users about
the changed behavior (resetting vs. deletion).
- Add the AttrDict.reset_field method to make available the original
AttrDict.__delitem__ behavior.
- Utilize typing.Annotated and pydantic markers to further simplify
serialization of pydantic.types.SecretStr fields in SecretConfig.
- Utilize typing.Annotated and pydantic markers to add exclude=True
parameter to token, client_id, and client_secret in TripleStoreConfig to
exclude these fields when serializing the model. This reverts back to
the original behavior for this model prior to upgrading to pydantic v2,
which is excellent!
- Add more explanatory comments where indicated in the review.
- Highlight change in serializing to JSON from a pydantic model when
comparing to json.dumps() (used only in tests).
- Update tests according to changes in the code base, attempting to keep
as much of the existing unit tests as possible.

Co-authored-by: Daniel Marchand <Daniel.Marchand@sintef.no>