Ensure config models are dumped safely (#443)
When creating a new strategy, if the given configuration is a pydantic
model it is dumped/serialized as a Python dict before it is provided to
the strategy upon initialization.
To ensure only the _actually_ provided configuration entries are passed
on `exclude_unset=True` should be set. And furthermore, to ensure all
values are valid JSON - something it should indeed already be if used
through OTEAPI Services - the dump mode has been set to `"json"`.
This will essentially ensure all values and nested values in the dumped
model are valid JSON types.